# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Stellabill Backend API
  description: |
    Stellabill backend HTTP API.

    This specification covers all implemented `/api/*` routes.

    Notes:
      - Preflight CORS requests (`OPTIONS`) are handled by middleware and return `204`.
  version: 0.2.0
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/subscriptions:
    get:
      tags: [Subscriptions]
      summary: List subscriptions
      operationId: listSubscriptions
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
          description: Pagination cursor for the next page of results
          example: "Y3Vyc29yX25leHRfcGFnZQ=="
        - name: limit
          in: query
          schema:
            type: integer
            maximum: 100
          description: Maximum number of items to return
          example: 20
      responses:
        "200":
          description: Subscriptions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
...
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [error, message, code]
      properties:
        error:
          type: string
          description: High-level error type
        message:
          type: string
          description: Human-readable error detail
        code:
          type: string
          description: Machine-readable error code
    Pagination:
      type: object
      additionalProperties: false
      required: [has_more]
      properties:
        next_cursor:
          type: string
          description: Opaque token to retrieve the next page of results
          example: "Y3Vyc29yX25leHRfcGFnZQ=="
        has_more:
          type: boolean
          description: Indicates if there are more results available
          example: true
...
    Subscription:
      type: object
      additionalProperties: false
      required: [id, plan_id, customer, status, amount, interval]
      properties:
        id:
          type: string
          example: sub_123
        plan_id:
          type: string
          example: plan_basic
        customer:
          type: string
          example: customer_123
        status:
          type: string
          example: active
        amount:
          type: string
          example: "1000"
        interval:
          type: string
          example: monthly
        next_billing:
          type: string
          example: "2026-04-01T00:00:00Z"
    SubscriptionsResponse:
      type: object
      additionalProperties: false
      required: [subscriptions, pagination]
      properties:
        subscriptions:
          type: array
          items:
            $ref: "#/components/schemas/Subscription"
        pagination:
          $ref: "#/components/schemas/Pagination"
  responses:
    BadRequest:
      description: Validation or client error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            PaginationLimitExceeded:
              value:
                error: "Bad Request"
                message: "Pagination limit cannot exceed 100 items."
                code: "validation_error"
            InvalidCursorFormat:
              value:
                error: "Bad Request"
                message: "Invalid pagination cursor format provided."
                code: "invalid_cursor"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            TokenMissing:
              value:
                error: "Unauthorized"
                message: "Authentication token is missing or invalid."
                code: "auth_unauthorized"
            TokenExpired:
              value:
                error: "Unauthorized"
                message: "Authentication token has expired."
                code: "auth_token_expired"
...